requirements.txt
streamlit
openai-whisper
transformers
PyPDF2
pydub
torch
sentencepiece
